Divided `CairoExtensions` among several files
